+++ src/components/sdk/MeetingControls/AudioInputVFControl.tsx
@@ -45,7 +45,7 @@ const AudioInputVFControl: React.FC<Props> = ({
   const meetingManager = useMeetingManager();
   const [isLoading, setIsLoading] = useState(false);
   // When the user click on Amazon Voice Focus option, the state will change.
-  const [isVoiceFocusOn, setIsVoiceFocusOn] = useState(false);
+  const [isVoiceFocusChecked, setIsVoiceFocusChecked] = useState(false);
   // Only when the current input audio device is an Amazon Voice Focus device, the state will be true. Otherwise, it will be false.
   const [isVoiceFocusEnabled, setIsVoiceFocusEnabled] = useState(false);
   const [dropdownWithVFOptions, setDropdownWithVFOptions] = useState<ReactNode[] | null>(null);
@@ -65,6 +65,10 @@ const AudioInputVFControl: React.FC<Props> = ({
   }, []);
 
   useEffect(() => {
+    console.info(`Amazon Voice Focus is ${isVoiceFocusEnabled ? 'enabled' : 'disabled'}.`);
+  }, [isVoiceFocusEnabled]);
+
+  useEffect(() => {
     // Only when the current input audio device is an Amazon Voice Focus transform device,
     // Amazon Voice Focus button will be checked.
     if (device instanceof VoiceFocusTransformDevice) {
@@ -81,7 +85,7 @@ const AudioInputVFControl: React.FC<Props> = ({
         children={<span>{device.label}</span>}
         checked={isOptionActive(selectedDevice, device.deviceId)}
         onClick={async (): Promise<void> => {
-          if (isVoiceFocusOn && !isLoading) {
+          if (isVoiceFocusChecked && !isLoading) {
             setIsLoading(true);
             const receivedDevice = audioInputSelectionToDevice(device.deviceId);
             const currentDevice = await addVoiceFocus(receivedDevice);
@@ -107,7 +111,7 @@ const AudioInputVFControl: React.FC<Props> = ({
           disabled={isLoading}
           onClick={() => {
             setIsLoading(true);
-            setIsVoiceFocusOn(current => !current);
+            setIsVoiceFocusChecked(current => !current);
             if (toggleVoiceFocus) {
               toggleVoiceFocus();
             }
@@ -128,7 +132,7 @@ const AudioInputVFControl: React.FC<Props> = ({
     devices[0].label,
     isLoading,
     isVoiceFocusEnabled,
-    isVoiceFocusOn,
+    isVoiceFocusChecked,
     isVoiceFocusSupported,
     selectedDevice,
   ]);
@@ -136,12 +140,14 @@ const AudioInputVFControl: React.FC<Props> = ({
   useEffect(() => {
     async function onVFCheckboxChange() {
       let current = device;
-      if (isVoiceFocusOn) {
+      if (isVoiceFocusChecked) {
+        console.info('User turned on Amazon Voice Focus.')
         if (typeof (device) === 'string') {
           const currentDevice = audioInputSelectionToDevice(device);
           current = await addVoiceFocus(currentDevice);
         }
       } else {
+        console.info('Amazon Voice Focus is off by default or user turned off Amazon Voice Focus.')
         if (device instanceof VoiceFocusTransformDevice) {
           current = device.getInnerDevice();
         }
@@ -151,7 +157,7 @@ const AudioInputVFControl: React.FC<Props> = ({
     }
 
     onVFCheckboxChange();
-  }, [isVoiceFocusOn]);
+  }, [isVoiceFocusChecked]);
 
   return (
     <ControlBarButton